+++ src/components/film-card/card-film.tsx
@@ -1,0 +1,19 @@
+type CardFilmProps = {
+  nameFilm: string;
+  imgPath: string;
+}
+
+function CardFilm(props: CardFilmProps) {
+  return (
+    <article className="small-film-card catalog__films-card">
+      <div className="small-film-card__image">
+        <img src={props.imgPath} alt={props.nameFilm} width="280" height="175"/>
+      </div>
+      <h3 className="small-film-card__title">
+        <a className="small-film-card__link" href="film-page.html">{props.nameFilm}</a>
+      </h3>
+    </article>
+  );
+}
+
+export default CardFilm;
