+++ src/components/film-card/film-card-text.tsx
@@ -1,0 +1,17 @@
+type FilmCardTextProps = {
+  description: string;
+  director: string;
+  starring: string;
+}
+
+function FilmCardText({description, director, starring}: FilmCardTextProps) {
+  return (
+    <div className="film-card__text">
+      {description}
+      <p className="film-card__director"><strong>Director: {director}</strong></p>
+      <p className="film-card__starring"><strong>Starring: {starring}</strong></p>
+    </div>
+  );
+}
+
+export default FilmCardText;
