+++ src/components/logo/logo.tsx
@@ -1,0 +1,17 @@
+import {Link} from 'react-router-dom';
+
+type LogoProps = {
+  className: string;
+}
+function Logo(props: LogoProps){
+  return(
+    <div className="logo">
+      <Link to="main.html" className={props.className}>
+        <span className="logo__letter logo__letter--1">W</span>
+        <span className="logo__letter logo__letter--2">T</span>
+        <span className="logo__letter logo__letter--3">W</span>
+      </Link>
+    </div>
+  );
+}
+export default Logo;
