--- src/components/user-block.tsx
@@ -1,20 +1,0 @@
-type UserBlockType = {
-  imgPath: string;
-}
-
-function UserBlock({imgPath}: UserBlockType) {
-  return (
-    <ul className="user-block">
-      <li className="user-block__item">
-        <div className="user-block__avatar">
-          <img src={imgPath} alt="User avatar" width="63" height="63"/>
-        </div>
-      </li>
-      <li className="user-block__item">
-        <a className="user-block__link">Sign out</a>
-      </li>
-    </ul>
-  );
-}
-
-export default UserBlock;
