+++ src/components/user-block/user-block.tsx
@@ -1,0 +1,20 @@
+type UserBlockProps = {
+  imgPath: string;
+}
+
+function UserBlock(props: UserBlockProps) {
+  return (
+    <ul className="user-block">
+      <li className="user-block__item">
+        <div className="user-block__avatar">
+          <img src={props.imgPath} alt="User avatar" width="63" height="63"/>
+        </div>
+      </li>
+      <li className="user-block__item">
+        <a className="user-block__link">Sign out</a>
+      </li>
+    </ul>
+  );
+}
+
+export default UserBlock;
