--- src/pages/head-guest-page.tsx
@@ -1,23 +1,0 @@
-import Logo from '../components/logo';
-
-function HeadGuestPage(){
-  return(
-    <section className="film-card">
-      <div className="film-card__bg">
-        <img src="img/bg-header.jpg" />
-      </div>
-
-      <h1 className="visually-hidden">WTW</h1>
-
-      <header className="page-header">
-        <Logo/>
-
-        <div className="user-block">
-          <a href="sign-in.html" className="user-block__link">Sign in</a>
-        </div>
-      </header>
-
-    </section>
-  );
-}
-export default HeadGuestPage;
