+++ src/pages/player/player-page.tsx
@@ -1,0 +1,9 @@
+import {Player} from '../../components/player/player.tsx';
+
+function PlayerPage() {
+  return (
+    <Player nameButton={'Play'} xlinkHref={'#play-s'}/>
+  );
+}
+
+export default PlayerPage;
