+++ src/pages/player/player-pause-page.tsx
@@ -1,0 +1,9 @@
+import {Player} from '../../components/player/player.tsx';
+
+function PlayerPausePage() {
+  return (
+    <Player nameButton={'Pause'} xlinkHref={'#pause'}/>
+  );
+}
+
+export default PlayerPausePage;
