+++ src/pages/sing-in-page/sing-in-page.tsx
@@ -1,0 +1,9 @@
+import {SingIn} from '../../components/sing-in/sing-in.tsx';
+
+function SingInPage() {
+  return (
+    <SingIn classNameEmail={'sign-in__field'}/>
+  );
+}
+
+export default SingInPage;
