+++ style/not-found-style/not-found-style.css
@@ -1,0 +1,20 @@
+.error-code {
+  margin-top: 15%;
+  font-size: 96px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.explanation-code-error {
+  font-size: 50px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.information-not-found {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
