+++ style/player-style/player-style.css
@@ -1,0 +1,3 @@
+.player__toggler {
+  left: 30%;
+}
